requirements.txt
influxdb
Pillow
gpiozero